questions/reconstruct_a_sequence.py
"""
Problem: Unique Reconstruction of Sequence

Description:
Given an original sequence 'originalSeq' and a list of sequences 'seqs', we need to determine whether 'originalSeq' can be uniquely reconstructed from 'seqs'. A unique reconstruction means that 'originalSeq' is the only sequence such that all sequences in 'seqs' are subsequences of it.

Function Signature:
def can_construct(originalSeq: List[int], seqs: List[List[int]]) -> bool:

Inputs:
    - originalSeq (List[int]): The main sequence which we want to check if it can be uniquely reconstructed.
    - seqs (List[List[int]]): A list of sequences.

Returns:
    - bool: True if 'originalSeq' can be uniquely reconstructed from 'seqs', otherwise False.

Examples:
1. Input: originalSeq: [1, 2, 3, 4], seqs: [[1, 2], [2, 3], [3, 4]]
   Output: True
   Explanation: The sequences [1, 2], [2, 3], and [3, 4] can uniquely reconstruct   
   [1, 2, 3, 4], implying that all the given sequences define the unique order of numbers in 'originalSeq'.

2. Input: originalSeq: [1, 2, 3, 4], seqs: [[1, 2], [2, 3], [2, 4]]
   Output: False
   Explanation: Sequences [1, 2], [2, 3], and [2, 4] cannot uniquely reconstruct [1, 2, 3, 4]. 
   There are two possible sequences derived from the given sequences: [1, 2, 3, 4] and [1, 2, 4, 3].

3. Input: originalSeq: [3, 1, 4, 2, 5], seqs: [[3, 1, 5], [1, 4, 2, 5]]
   Output: True
   Explanation: Sequences [3, 1, 5] and [1, 4, 2, 5] can uniquely reconstruct [3, 1, 4, 2, 5].

Notes:
    - This problem can be approached by leveraging the topological sort. We can consider each number as a unique node and draw directed edges from each sequence in 'seqs'.
    - If the resulting topological order is unique and matches the 'originalSeq', then the sequence can be uniquely reconstructed.

Tags:
    - Graphs
    - Topological Sorting
"""
from typing import List

def can_construct(originalSeq: List[int], seqs: List[List[int]]) -> bool:
  return False